Refatoração do MarcaController

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,10 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +18,9 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        return Marca::all();
+        //return Marca::all();
+        $marca = $this->marca->all();
+        return response()->json($marca, 200);
     }
 
     /**
@@ -25,42 +31,65 @@ class MarcaController extends Controller
      */
     public function store(Request $req)
     {
-        return Marca::create($req->all());
+        //return Marca::create($req->all());
+        $marca = $this->marca->create($req->all());
+
+        return response()->json($marca, 201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'NotFound'], 404);
+        }
+
+        return response()->json($marca, 200);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $req, Marca $marca)
+    public function update(Request $req, $id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'NotFound'], 404);
+        }
+
         $marca->update($req->all());
-        return $marca;
+
+        return response()->json($marca, 200);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'NotFound'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'sucesso'];
+
+        return response()->json(['msg' => 'sucesso'], 200);
     }
 }
